step1 refactor(bowling) NormalFrame Strike,Spare,Miss 로 분리
