+++ src/main/java/bowling/LastFallDownPins.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.state.State;
 import java.util.ArrayList;
 import java.util.List;
