+++ src/test/java/bowling/NormalFrameTest.java
@@ -2,6 +2,7 @@ package bowling;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import bowling.state.FirstRoll;
 import org.junit.jupiter.api.Test;
 
 public class NormalFrameTest {
